warning: SHUT UP!

+++ sigterm/src/audio/AudioConverter.cc
@@ -36,7 +36,7 @@ bool AudioConverter::convert(AudioBuffer *inOutAudioBuffer) {
     }
 
     //qDebug("converter: converted to %d (%d * %f) bytes", (int)(mCVT.len * mCVT.len_ratio), mCVT.len, mCVT.len_ratio);
-    inOutAudioBuffer->setConvertedChunkLength(mCVT.len * mCVT.len_ratio);
+    inOutAudioBuffer->setConvertedChunkLength((quint32)(mCVT.len * mCVT.len_ratio));
 
     return true;
 }
